Refactor project config and web host helpers
